INV-9: Fixed mismatched data type on JPA receipt repository

+++ src/main/java/com/invoicemanager/invoice_manager/receipt/repository/ReceiptRepository.java
@@ -5,9 +5,9 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.invoicemanager.invoice_manager.receipt.model.Receipt;
 
-public interface ReceiptRepository extends JpaRepository<Receipt, Integer> {
+public interface ReceiptRepository extends JpaRepository<Receipt, Long> {
 
   Optional<Receipt> findByReceiptNumber(String receiptNumber);
 
-  List<Receipt> findByCustomerId(Integer customerId);
+  List<Receipt> findByCustomerId(Long customerId);
 }
